BIGUQ.jl

BIGUQ.jl implements Bayesian Information Gap Decision Theory (BIGDT) for Uncertainty Quantification, Experimental Design and Decision Analysis.

Relevant examples:

Information Gap Analysis

Decision Analysis

Model Diagnostics

BIGUQ.jl functions:

BIGUQ.getmcmcchain — Method

Get MCMC chain

BIGUQ.makebigdts — Method

Make BIGDT analyses for each possible decision assuming that the proposed observations proposedobs are observed

BIGUQ.makebigdts — Method

Makes BIGDT analyses for each possible decision assuming that no more observations will be made

BIGUQ.test — Method

Test BIGUQ functions

BIGUQ.BigDT — Type

BigOED type

BIGUQ.BigOED — Type

BigOED type

References:

O’Malley, D., Vesselinov, V.V., A combined probabilistic/non-probabilistic decision analysis for contaminant remediation, Journal on Uncertainty Quantification, SIAM/ASA, 10.1137/140965132, 2014.

O’Malley, D., Vesselinov, V.V., Bayesian-Information-Gap decision theory with an application to CO2 sequestration, Water Resources Research, 10.1002/2015WR017413, 2015.

Grasinger, M., O'Malley, D., Vesselinov, V.V., Karra, S., Decision Analysis for Robust CO2 Injection: Application of Bayesian-Information-Gap Decision Theory, International Journal of Greenhouse Gas Control, 10.1016/j.ijggc.2016.02.017, 2016.