ReusableFunctions.jl

ReusableFunctions.jl provides automated storage and retrieval of results for Julia function calls.

ReusableFunctions.jl helps avoiding repetative computations for same function call inputs.

ReusableFunctions.jl supports efficient restarts and reuse of past computations.

ReusableFunctions.jl functions:

ReusableFunctions.checkhashfilename — Method

Check if a file with a filename based on hash exists

ReusableFunctions.gethashfilename — Method

Define a filename based on hash

ReusableFunctions.loadresultfile — Method

Load JLD result file

ReusableFunctions.maker3function — Method

Make a reusable function expecting both regular and keyword arguments

ReusableFunctions.quietoff — Method

Make ReusableFunctions not quiet

ReusableFunctions.quieton — Method

Make ReusableFunctions quiet

ReusableFunctions.resetcomputes — Method

Reset computes counter

ReusableFunctions.resetrestarts — Method

Reset restarts counter

ReusableFunctions.saveresultfile — Method

Save JLD result file

ReusableFunctions.test — Method

Test ReusableFunctions